pyproject.toml
[project]
name = "ebook-audiobook"
version = "0.1.0"
description = ""
authors = [
    {name = "MKHAZZANI",email = "m.khazzani2001@gmail.com"}
]
readme = "README.md"
packages = [{include = "src"}]
requires-python = ">=3.11,<3.12"
dependencies = [
    "click (>=8.2.1,<9.0.0)",
    "gradio (>=5.39.0,<6.0.0)",
    "django (>=5.0.7,<6.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "tts (>=0.22.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
